@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ArtistsController, type: :controller do
   # This should return the minimal set of values that should be in the session
@@ -9,7 +9,7 @@ describe ArtistsController, type: :controller do
   describe 'GET index' do
     it 'assigns all artists as @artists' do
       artist = create(:artist)
-      get :index, {}, valid_session
+      get :index, params: {}, session: valid_session
       expect(assigns(:artists)).to eq([artist])
     end
   end
@@ -17,7 +17,7 @@ describe ArtistsController, type: :controller do
   describe 'GET show' do
     it 'assigns the requested artist as @artist' do
       artist = create(:artist)
-      get :show, {id: artist.to_param}, valid_session
+      get :show, params: { id: artist.to_param }, session: valid_session
       expect(assigns(:artist)).to eq(artist)
     end
   end
@@ -27,13 +27,21 @@ describe ArtistsController, type: :controller do
     describe 'with valid params' do
       it 'updates the requested artist' do
         artist = create(:artist)
-        expect_any_instance_of(Artist).to receive(:update).with({ 'rating' => '1'})
-        put :update, {id: artist.to_param, artist: {'rating' => '1'}, format: :json}, valid_session
+        expect_any_instance_of(Artist).to receive(:update).with(
+            ActionController::Parameters.new({ 'rating' => '1'}).permit(:rating)
+        )
+        put :update, params: { id: artist.to_param,
+                               artist: { 'rating' => '1' },
+                               format: :json },
+            session: valid_session
       end
 
       it 'assigns the requested artist as @artist' do
         artist = create(:artist)
-        put :update, {id: artist.to_param, artist: attributes_for(:artist), format: :json}, valid_session
+        put :update, params: { id: artist.to_param,
+                               artist: attributes_for(:artist),
+                               format: :json },
+            session: valid_session
         expect(assigns(:artist)).to eq(artist)
       end
     end
@@ -43,7 +51,10 @@ describe ArtistsController, type: :controller do
         artist = create(:artist)
         # Trigger the behavior that occurs when invalid params are submitted
         allow_any_instance_of(Artist).to receive(:update).and_return(false)
-        put :update, {id: artist.to_param, artist: {'rating' => '-2'}, format: :json}, valid_session
+        put :update, params: { id: artist.to_param,
+                               artist: { 'rating' => '-2' },
+                               format: :json },
+            session: valid_session
         expect(assigns(:artist)).to eq(artist)
       end
     end
