@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 # This spec was generated by rspec-rails when you ran the scaffold generator.
 # It demonstrates how one might use RSpec to specify the controller code that
@@ -27,7 +27,7 @@ describe GenresController, type: :controller do
   describe 'GET index' do
     it 'assigns all genres as @genres' do
       genre = create(:genre)
-      get :index, {}, valid_session
+      get :index, params: {}, session: valid_session
       expect(assigns(:genres)).to eq([genre])
     end
   end
@@ -35,7 +35,7 @@ describe GenresController, type: :controller do
   describe 'GET show' do
     it 'assigns the requested genre as @genre' do
       genre = create(:genre)
-      get :show, {id: genre.to_param}, valid_session
+      get :show, params: { id: genre.to_param }, session: valid_session
       expect(assigns(:genre)).to eq(genre)
     end
   end
