@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 # This spec was generated by rspec-rails when you ran the scaffold generator.
 # It demonstrates how one might use RSpec to specify the controller code that
@@ -27,7 +27,7 @@ RSpec.describe SettingsController, type: :controller do
   describe 'GET index' do
     it 'assigns all setting as @settings' do
       setting = create(:settings)
-      get :index, {format: :json}, valid_session
+      get :index, params: { format: :json }, session: valid_session
       expect(assigns(:settings)).to eq(Settings.unscoped.get_all)
     end
   end
@@ -35,7 +35,7 @@ RSpec.describe SettingsController, type: :controller do
   describe 'GET show' do
     it 'assigns the requested artist as @setting' do
       setting = create(:settings)
-      get :show, {var: setting.var, format: :json}, valid_session
+      get :show, params: { var: setting.var, format: :json }, session: valid_session
       expect(assigns(:setting)).to eq(setting.value)
     end
   end
@@ -44,7 +44,8 @@ RSpec.describe SettingsController, type: :controller do
     describe 'with valid params' do
       it 'creates a new Setting' do
         expect {
-          post(:create, {var:'uuid2', value:'4567'}, valid_session)
+          post :create, params: { var:'uuid2', value:'4567' },
+               session: valid_session
         }.to change(Settings, :count).by(1)
       end
     end
@@ -58,7 +59,8 @@ RSpec.describe SettingsController, type: :controller do
         # receives the :update_attributes message with whatever params are
         # submitted in the request.
         settings = create(:settings)
-        put :update, {var: 'uuid', value: '8910', format: :json}, valid_session
+        put :update, params: { var: 'uuid', value: '8910', format: :json },
+            session: valid_session
         expect(Settings[:uuid]).to eql('8910')
       end
     end
@@ -67,7 +69,7 @@ RSpec.describe SettingsController, type: :controller do
     it 'destroys the requested setting' do
       settings = create(:settings, var:'uuid3', value:'abcd')
       expect {
-        delete :destroy, {var:'uuid3'}, valid_session
+        delete :destroy, params: { var:'uuid3' }, session: valid_session
       }.to change(Settings, :count).by(-1)
     end
   end
