@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 # This spec was generated by rspec-rails when you ran the scaffold generator.
 # It demonstrates how one might use RSpec to specify the controller code that
@@ -31,14 +31,14 @@ describe TracksController, type: :controller do
 
   describe 'GET index' do
     it 'assigns all tracks as @tracks' do
-      get :index, {}, valid_session
+      get :index, params: {}, session: valid_session
       expect(assigns(:tracks)).to eq([@track])
     end
   end
 
   describe 'GET show' do
     it 'assigns the requested track as @track' do
-      get :show, {id: @track.to_param}, valid_session
+      get :show, params: { id: @track.to_param }, session: valid_session
       expect(assigns(:track)).to eq(@track)
     end
   end
@@ -47,35 +47,35 @@ describe TracksController, type: :controller do
     describe 'with valid params' do
       it 'creates a new Track' do
         expect {
-          post(:create, {track: attributes_for(:track,
-                                               title: 'Tokyo Skies',
-                                               track_id: 2,
-                                               location: '/mnt/Data/Music/Chipzel/Spectra/Tokyo Skies.flac'),
-                                album: @track.album.title,
-                                artist: @track.artist.name, format: :json },
-               valid_session)
+          post :create, params: {
+              track: attributes_for(:track, title: 'Tokyo Skies',
+                                    track_id: 2, location: '/mnt/Data/Music/Chipzel/Spectra/Tokyo Skies.flac'),
+              album: @track.album.title,
+              artist: @track.artist.name, format: :json
+          }, session: valid_session
         }.to change(Track, :count).by(1)
       end
 
       it 'assigns a newly created track as @track' do
-        post(:create, {track: attributes_for(:track,
-                                             title: 'Forged in Stars',
-                                             track_id: 3,
-                                             location: '/mnt/Data/Music/Chipzel/Spectra/Forged in Stars.flac'),
-                      album: @track.album.title,
-                      artist: @track.artist.name, format: :json}, valid_session)
+        post :create, params: {
+            track: attributes_for(:track, title: 'Forged in Stars',
+                                  track_id: 3, location: '/mnt/Data/Music/Chipzel/Spectra/Forged in Stars.flac'),
+            album: @track.album.title,
+            artist: @track.artist.name,
+            format: :json
+        }, session: valid_session
         expect(assigns(:track)).to be_a(Track)
         expect(assigns(:track)).to be_persisted
       end
 
       it 'redirects to the created track' do
-        post(:create, {track: attributes_for(:track,
-                                             title: 'Formed in the Clouds',
-                                             track_id: 4,
-                                             location: '/mnt/Data/Music/Chipzel/Spectra/Formed in the Clouds.flac'),
-                              album: @track.album.title,
-                      artist: @track.artist.name, format: :json},
-             valid_session)
+        post :create, params: {
+            track: attributes_for(:track, title: 'Formed in the Clouds',
+                                  track_id: 4, location: '/mnt/Data/Music/Chipzel/Spectra/Formed in the Clouds.flac'),
+            album: @track.album.title,
+            artist: @track.artist.name,
+            format: :json
+        }, session: valid_session
         track = Track.last
         expect(response).to have_http_status(:created)
       end
@@ -85,14 +85,14 @@ describe TracksController, type: :controller do
       it 'assigns a newly created but unsaved track as @track' do
         # Trigger the behavior that occurs when invalid params are submitted
         allow_any_instance_of(Track).to receive(:save).and_return(false)
-        post :create, { track: {'title' => ''}, format: :json}, valid_session
+        post :create, params: { track: { 'title' => '' }, format: :json}, session: valid_session
         expect(assigns(:track)).to be_a_new(Track)
       end
 
       it "re-renders the 'new' template" do
         # Trigger the behavior that occurs when invalid params are submitted
         allow_any_instance_of(Track).to receive(:save).and_return(false)
-        post :create, {track: {'title' => ''}, format: :json}, valid_session
+        post :create, params: { track: { 'title' => '' }, format: :json }, session: valid_session
         expect(response).to be_unprocessable
       end
     end
@@ -105,20 +105,32 @@ describe TracksController, type: :controller do
         # specifies that the Track created on the previous line
         # receives the :update_attributes message with whatever params are
         # submitted in the request.
-        expect_any_instance_of(Track).to receive(:update).with({'title' => 'Only Human', 'track_id' => '5'})
-        put :update, {id: @track.to_param,
-                      track: {'title' => 'Only Human', 'track_id' => '5'}, format: :json}, valid_session
+        expect_any_instance_of(Track).to receive(:update).with(
+            ActionController::Parameters.new({ 'title' => 'Only Human', 'track_id' => '5' }).permit(:title, :track_id)
+        )
+
+        put :update, params: {
+            id: @track.to_param,
+            track: { 'title' => 'Only Human', 'track_id' => '5' },
+            format: :json
+        }, session: valid_session
       end
 
       it 'assigns the requested track as @track' do
-        put :update, {id: @track.to_param,
-                      track: attributes_for(:track), format: :json}, valid_session
+        put :update, params: {
+            id: @track.to_param,
+            track: attributes_for(:track),
+            format: :json
+        }, session: valid_session
         expect(assigns(:track)).to eq(@track)
       end
 
       it 'redirects to the track' do
-        put :update, {id: @track.to_param,
-                      track: attributes_for(:track), format: :json}, valid_session
+        put :update, params: {
+            id: @track.to_param,
+            track: attributes_for(:track),
+            format: :json
+        }, session: valid_session
         expect(response).to be_ok
       end
     end
@@ -127,16 +139,22 @@ describe TracksController, type: :controller do
       it 'assigns the track as @track' do
         # Trigger the behavior that occurs when invalid params are submitted
         allow_any_instance_of(Track).to receive(:update).and_return(false)
-        put :update, {id: @track.to_param,
-                      track: {'title' => 'invalid value'}, format: :json}, valid_session
+        put :update, params: {
+            id: @track.to_param,
+            track: { 'title' => 'invalid value' },
+            format: :json
+        }, session: valid_session
         expect(assigns(:track)).to eq(@track)
       end
 
       it "re-renders the 'edit' template" do
         # Trigger the behavior that occurs when invalid params are submitted
         allow_any_instance_of(Track).to receive(:update).and_return(false)
-        put :update, {id: @track.to_param,
-                      track: {'title' => 'invalid value'}, format: :json}, valid_session
+        put :update, params: {
+            id: @track.to_param,
+            track: { 'title' => 'invalid value' },
+            format: :json
+        }, session: valid_session
         expect(response).to be_unprocessable
       end
     end
@@ -145,7 +163,7 @@ describe TracksController, type: :controller do
   describe 'DELETE destroy' do
     it 'destroys the requested track' do
       expect {
-        delete :destroy, {id: @track.to_param, format: :json}, valid_session
+        delete :destroy, params: { id: @track.to_param, format: :json }, session: valid_session
       }.to change(Track, :count).by(-1)
     end
 
