@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe 'Albums', type: :request do
   describe 'GET /albums' do
